requirements.txt
requests
google-cloud-pubsub
python-dateutil
functions-framework